Boves Ceramics

A dataset containing the ceramic counts from the castle site of Boves (Somme, France). The data are grouped into eight periods ranging from the 10th to the 18th century and thirteen ceramic types.

boves

Format

A data frame with 8 rows and 13 variables (ceramic types):

I

IIa

IIb

IIc

IIIa

IIIb

IIIc

IVa

IVb

Va

Vb

VI

VII

Source

Racinet P. (2002). Le site castral et prioral de Boves du Xe au XVIIe siècle. Bilan des recherches 1996-2000. Revue archéologique de Picardie. Numéro spécial 20, 123 p.

See also