Compiègne Ceramics

A dataset containing the ceramic counts from the Place des Hallettes in Compiègne (Oise, France). The data are grouped into five periods of about a century, ranging from the 9th to the 14th century, and sixteen ceramic types.

Usage

compiegne

Format

A data.frame with 5 rows (chronological periods, numbered from the oldest to the most recent from 1 to 5) and 16 variables (ceramic types):

A

Red to white ceramics with fine sized inclusions.

B

Red to white ceramics with medium sized inclusions.

C

Dark ceramics with fine sized inclusions.

D

Dark ceramics with medium sized inclusions.

E

Ceramics close to those of groups B or D, with similarities to group F.

F

Black, red or beige ceramics with coarse inclusions.

G

Red polished ceramics with fine to medium sized inclusions.

H

Black polished ceramics with fine sized inclusions.

I

Black polished ceramics with medium sized inclusions.

J

Polished and painted ceramics with fine to medium sized inclusions.

K

Painted ceramics, similar to those of group A.

L

Painted ceramics, similar to those of group B.

M

Painted ceramics with coarse inclusions.

N

Glazed ceramics.

O

Stamped ceramics.

P

Coated ceramics.

Source

Lacroix, M. C. (1997). La céramique médiévale du site des Hallettes à Compiègne (Oise). Revue archéologique de Picardie. Numéro spécial, 13(1), 135-168. doi:10.3406/pica.1997.1945

See also